# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: КТРУ-классификатор API
  description: Сопоставление позиций спецификаций с кодами КТРУ
  version: 0.1.0
paths:
  /api/v1/health:
    get:
      tags:
      - health
      summary: Health
      operationId: health_api_v1_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /api/v1/jobs:
    post:
      tags:
      - jobs
      summary: Create Job
      operationId: create_job_api_v1_jobs_post
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_create_job_api_v1_jobs_post'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreated'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - jobs
      summary: List Jobs
      operationId: list_jobs_api_v1_jobs_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/{job_id}:
    get:
      tags:
      - jobs
      summary: Get Job
      operationId: get_job_api_v1_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/{job_id}/items:
    get:
      tags:
      - jobs
      summary: List Items
      operationId: list_items_api_v1_jobs__job_id__items_get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Job Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Status
      - name: min_confidence
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            maximum: 100
            minimum: 0
          - type: 'null'
          title: Min Confidence
      - name: max_confidence
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            maximum: 100
            minimum: 0
          - type: 'null'
          title: Max Confidence
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 256
          - type: 'null'
          title: Q
      - name: sort_by
        in: query
        required: false
        schema:
          enum:
          - row_number
          - confidence
          - updated_at
          type: string
          default: row_number
          title: Sort By
      - name: sort_order
        in: query
        required: false
        schema:
          enum:
          - asc
          - desc
          type: string
          default: asc
          title: Sort Order
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/{job_id}/accept:
    post:
      tags:
      - jobs
      summary: Accept Items
      operationId: accept_items_api_v1_jobs__job_id__accept_post
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Job Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkAccept'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkAcceptResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/{job_id}/export:
    get:
      tags:
      - jobs
      summary: Export Job
      operationId: export_job_api_v1_jobs__job_id__export_get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Job Id
      - name: format
        in: query
        required: false
        schema:
          enum:
          - csv
          - xlsx
          type: string
          default: csv
          title: Format
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/items/{item_id}:
    patch:
      tags:
      - items
      summary: Correct Item
      operationId: correct_item_api_v1_items__item_id__patch
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCorrection'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/items/{item_id}/accept:
    post:
      tags:
      - items
      summary: Accept Item
      operationId: accept_item_api_v1_items__item_id__accept_post
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/items/{item_id}/revert:
    post:
      tags:
      - items
      summary: Revert Item
      operationId: revert_item_api_v1_items__item_id__revert_post
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/ktru/search:
    get:
      tags:
      - catalog
      summary: Search Catalog
      operationId: search_catalog_api_v1_ktru_search_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 2
          maxLength: 256
          title: Q
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          default: 10
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KtruSearchPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/ktru/{ktru_code}:
    get:
      tags:
      - catalog
      summary: Get Position
      operationId: get_position_api_v1_ktru__ktru_code__get
      parameters:
      - name: ktru_code
        in: path
        required: true
        schema:
          type: string
          title: Ktru Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KtruPositionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Body_create_job_api_v1_jobs_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_create_job_api_v1_jobs_post
    BulkAccept:
      properties:
        min_confidence:
          type: number
          maximum: 100.0
          minimum: 0.0
          title: Min Confidence
          description: Принять все позиции с уверенностью не ниже
      type: object
      required:
      - min_confidence
      title: BulkAccept
      example:
        min_confidence: 85
    BulkAcceptResult:
      properties:
        accepted_count:
          type: integer
          title: Accepted Count
      type: object
      required:
      - accepted_count
      title: BulkAcceptResult
    CandidateOut:
      properties:
        ktru_code:
          type: string
          title: Ktru Code
        ktru_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Ktru Name
        confidence:
          type: number
          title: Confidence
      type: object
      required:
      - ktru_code
      - confidence
      title: CandidateOut
    Characteristic:
      properties:
        name:
          type: string
          title: Name
        value:
          type: string
          title: Value
        unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit
      type: object
      required:
      - name
      - value
      title: Characteristic
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HealthComponent:
      properties:
        name:
          type: string
          title: Name
        ok:
          type: boolean
          title: Ok
        detail:
          type: string
          title: Detail
      type: object
      required:
      - name
      - ok
      - detail
      title: HealthComponent
    HealthResponse:
      properties:
        status:
          type: string
          title: Status
        version:
          type: string
          title: Version
        components:
          items:
            $ref: '#/components/schemas/HealthComponent'
          type: array
          title: Components
      type: object
      required:
      - status
      - version
      - components
      title: HealthResponse
    ItemAttributes:
      properties:
        name:
          type: string
          title: Name
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        characteristics:
          items:
            $ref: '#/components/schemas/Characteristic'
          type: array
          title: Characteristics
      type: object
      required:
      - name
      title: ItemAttributes
    ItemCorrection:
      properties:
        ktru_code:
          type: string
          maxLength: 32
          minLength: 1
          title: Ktru Code
        ktru_name:
          anyOf:
          - type: string
            maxLength: 512
          - type: 'null'
          title: Ktru Name
      type: object
      required:
      - ktru_code
      title: ItemCorrection
    ItemOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        job_id:
          type: string
          format: uuid
          title: Job Id
        row_number:
          type: integer
          title: Row Number
        raw_name:
          type: string
          title: Raw Name
        raw_columns:
          additionalProperties: true
          type: object
          title: Raw Columns
        attributes:
          anyOf:
          - $ref: '#/components/schemas/ItemAttributes'
          - type: 'null'
        ktru_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Ktru Code
        ktru_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Ktru Name
        confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence
        method:
          anyOf:
          - type: string
          - type: 'null'
          title: Method
        status:
          type: string
          enum:
          - pending
          - auto
          - needs_review
          - accepted
          - corrected
          title: Status
        candidates:
          items:
            $ref: '#/components/schemas/CandidateOut'
          type: array
          title: Candidates
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - job_id
      - row_number
      - raw_name
      - raw_columns
      - ktru_code
      - ktru_name
      - confidence
      - method
      - status
      - updated_at
      title: ItemOut
      example:
        attributes:
          brand: Dell
          characteristics:
          - name: ОЗУ
            unit: ГБ
            value: '16'
          model: Latitude 5540
          name: Ноутбук
        candidates:
        - confidence: 94.0
          ktru_code: 26.20.11.110-00000009
          ktru_name: Машины вычислительные портативные
        confidence: 94.0
        id: 01aaa595-b008-4229-8a74-c3ac387f0261
        job_id: c82cc353-9233-4ba2-8270-4f0139f37774
        ktru_code: 26.20.11.110-00000009
        ktru_name: Машины вычислительные портативные
        method: hybrid
        raw_columns:
          Наименование товара: Ноутбук Dell
          №: '1'
        raw_name: Ноутбук Dell Latitude 5540 16 ГБ
        row_number: 2
        status: auto
        updated_at: '2026-10-05T15:41:27.812874Z'
    ItemPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ItemOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: ItemPage
    JobCreated:
      properties:
        job_id:
          type: string
          format: uuid
          title: Job Id
        status:
          type: string
          enum:
          - queued
          - processing
          - done
          - failed
          title: Status
        filename:
          type: string
          title: Filename
      type: object
      required:
      - job_id
      - status
      - filename
      title: JobCreated
      example:
        filename: spec.csv
        job_id: c82cc353-9233-4ba2-8270-4f0139f37774
        status: queued
    JobOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        filename:
          type: string
          title: Filename
        status:
          type: string
          enum:
          - queued
          - processing
          - done
          - failed
          title: Status
        total_count:
          type: integer
          title: Total Count
        processed_count:
          type: integer
          title: Processed Count
        failed_count:
          type: integer
          title: Failed Count
        auto_count:
          type: integer
          title: Auto Count
          default: 0
        needs_review_count:
          type: integer
          title: Needs Review Count
          default: 0
        corrected_count:
          type: integer
          title: Corrected Count
          default: 0
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        parse_seconds:
          anyOf:
          - type: number
          - type: 'null'
          title: Parse Seconds
        created_at:
          type: string
          format: date-time
          title: Created At
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
      type: object
      required:
      - id
      - filename
      - status
      - total_count
      - processed_count
      - failed_count
      - error
      - parse_seconds
      - created_at
      - started_at
      - finished_at
      title: JobOut
      example:
        auto_count: 3
        corrected_count: 0
        created_at: '2026-10-05T15:41:27.454156Z'
        failed_count: 0
        filename: spec.csv
        finished_at: '2026-10-05T15:41:27.817763Z'
        id: c82cc353-9233-4ba2-8270-4f0139f37774
        needs_review_count: 1
        parse_seconds: 0.019
        processed_count: 4
        started_at: '2026-10-05T15:41:27.792843Z'
        status: done
        total_count: 4
    JobPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/JobOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: JobPage
    KtruPositionOut:
      properties:
        ktru_code:
          type: string
          title: Ktru Code
        ktru_name:
          type: string
          title: Ktru Name
        okpd2_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Okpd2 Code
        characteristics:
          items:
            $ref: '#/components/schemas/Characteristic'
          type: array
          title: Characteristics
      type: object
      required:
      - ktru_code
      - ktru_name
      title: KtruPositionOut
      example:
        characteristics:
        - name: масса
          unit: кг
          value: не более 10
        ktru_code: 26.20.11.110-00000009
        ktru_name: Машины вычислительные портативные
        okpd2_code: 26.20.11.110
    KtruSearchPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/KtruPositionOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - items
      - total
      - limit
      title: KtruSearchPage
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError